src/BE/requirement.txt
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bitsandbytes==0.35.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
diffusers @ git+https://github.com/ShivamShrirao/diffusers@2c298300756fa7d8bf644852cffaebc5072f11f6
exceptiongroup==1.1.2
fastapi==0.99.1
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
gradio==3.35.2
gradio_client==0.2.7
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.2
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
linkify-it-py==2.0.2
lit==16.0.6
mangum==0.17.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
natsort==8.4.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
orjson==3.9.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
psutil==5.9.5
pydantic==1.10.11
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rpds-py==0.8.8
safetensors==0.3.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
websockets==11.0.3
xformers==0.0.20
yarl==1.9.2
zipp==3.15.0
